// packages/user-connectrpc-server/src/generated/chat3_user_pb.d.ts
export declare type GetDialogMessagesRequest = Message$1<"chat3.user.GetDialogMessagesRequest"> & {
  /**
   * @generated from field: string dialog_id = 1;
   */
  dialogId: string;

  /**
   * @generated from field: int32 page = 2;
   */
  page: number;

  /**
   * @generated from field: int32 limit = 3;
   */
  limit: number;

  /**
   * @generated from field: string filter = 4;
   */
  filter: string;

  /**
   * @generated from field: string sort = 5;
   */
  sort: string;
};

/**
 * Describes the message chat3.user.GetDialogMessagesRequest.
 * Use `create(GetDialogMessagesRequestSchema)` to create a new message.
 */
export declare const GetDialogMessagesRequestSchema: GenMessage<GetDialogMessagesRequest>;

/**
 * Ответ с сообщениями диалога
 *
 * @generated from message chat3.user.GetDialogMessagesResponse
 */
export declare type GetDialogMessagesResponse = Message$1<"chat3.user.GetDialogMessagesResponse"> & {
  /**
   * @generated from field: repeated chat3.user.Message messages = 1;
   */
  messages: Message[];

  /**
   * @generated from field: chat3.user.Pagination pagination = 2;
   */
  pagination?: Pagination;
};
// ...
export declare type Message = Message$1<"chat3.user.Message"> & {
  /**
   * @generated from field: string message_id = 1;
   */
  messageId: string;

  /**
   * @generated from field: string dialog_id = 2;
   */
  dialogId: string;

  /**
   * @generated from field: string sender_id = 3;
   */
  senderId: string;

  /**
   * @generated from field: string type = 4;
   */
  type: string;

  /**
   * @generated from field: string content = 5;
   */
  content: string;

  /**
   * @generated from field: google.protobuf.Struct meta = 6;
   */
  meta?: JsonObject;

  /**
   * @generated from field: repeated chat3.user.MessageStatus statuses = 7;
   */
  statuses: MessageStatus[];

  /**
   * @generated from field: google.protobuf.Struct reaction_set = 8;
   */
  reactionSet?: JsonObject;

  /**
   * @generated from field: chat3.user.SenderInfo sender_info = 9;
   */
  senderInfo?: SenderInfo;

  /**
   * @generated from field: double created_at = 10;
   */
  createdAt: number;

  /**
   * @generated from field: string topic_id = 11;
   */
  topicId: string;

  /**
   * @generated from field: google.protobuf.Struct topic = 12;
   */
  topic?: JsonObject;
};

/**
 * Describes the message chat3.user.Message.
 * Use `create(MessageSchema)` to create a new message.
 */
export declare const MessageSchema: GenMessage<Message>;

/**
 * Статус сообщения
 *
 * @generated from message chat3.user.MessageStatus
 */
export declare type MessageStatus = Message$1<"chat3.user.MessageStatus"> & {
  /**
   * @generated from field: string user_id = 1;
   */
  userId: string;

  /**
   * @generated from field: string status = 2;
   */
  status: string;

  /**
   * @generated from field: double read_at = 3;
   */
  readAt: number;

  /**
   * @generated from field: double created_at = 4;
   */
  createdAt: number;
};

/**
 * Describes the message chat3.user.MessageStatus.
 * Use `create(MessageStatusSchema)` to create a new message.
 */
export declare const MessageStatusSchema: GenMessage<MessageStatus>;

/**
 * Информация об отправителе
 *
 * @generated from message chat3.user.SenderInfo
 */
export declare type SenderInfo = Message$1<"chat3.user.SenderInfo"> & {
  /**
   * @generated from field: string user_id = 1;
   */
  userId: string;

  /**
   * @generated from field: string name = 2;
   */
  name: string;

  /**
   * @generated from field: double created_at = 3;
   */
  createdAt: number;

  /**
   * @generated from field: google.protobuf.Struct meta = 4;
   */
  meta?: JsonObject;
};

/**
 * Describes the message chat3.user.SenderInfo.
 * Use `create(SenderInfoSchema)` to create a new message.
 */
export declare const SenderInfoSchema: GenMessage<SenderInfo>;

/**
 * Пагинация
 *
 * @generated from message chat3.user.Pagination
 */
export declare type Pagination = Message$1<"chat3.user.Pagination"> & {
  /**
   * @generated from field: int32 page = 1;
   */
  page: number;

  /**
   * @generated from field: int32 limit = 2;
   */
  limit: number;

  /**
   * @generated from field: int32 total = 3;
   */
  total: number;

  /**
   * @generated from field: int32 pages = 4;
   */
  pages: number;
};